Fixed bug:
After logging out from a profile page, urlbar still shows profile_url
so that when logging in, you default to seeing the last displayed page
instead of your feed. Fix was to chain on to the promise in the logout
action and browserHistory push to /.
Maybe should do then when logging in too?